Implement token schedule based on global step
This is the first step towards implementing a k-synchronous
system in which we control the degree of synchronization through
the number of tokens we add to the queue in each step.

The hard-coded schedule doesn't actually mean anything yet,
and what is implemented is currently not a correct transition
between async and sync modes.