Refactor: separate autoscaling agent from hook
This refactor anticipates integrating autoscaling with the keras
API, which no longer uses sessions. This means we can no longer
use SessionRunHook to implement our agent. This commit abstracts
most of the functionality of the AutoscalingHook into a new class
called AutoscalingAgent, which will be shared by the integration
with the keras API.