Fix race condition with interleaved join and sync
Here's how the race condition happens:
1. Worker X finishes syncing cluster spec
2. New worker Y joins
3. Worker X *overwrites* its own cluster_spec while handling
   the join request from worker Y
4. Worker X sets TF_CONFIG using the overwritten cluster_spec

The overwriting behavior in step 3 certainly is not intended.
When handling an add_worker (or remove_worker) request in
AutoscalingService, we actually only intend to modify
self.agent.pending_cluster_spec, but actually we also modify
self.agent.cluster_spec. This is because self.get_cluster_spec
actually returns a pointer to self.agent.cluster_spec instead
of a copy.

Thus, the fix is simple: simply return a deep copy of
self.agent.cluster_spec instead of the object.