Run the correct number of steps even with restarts
The key idea is the number of steps to run is the same no matter
how many workers we have. This means that we just need to keep
track of the number of steps we've run so far.

The way this is implemented is as follows: if we restarted in the
middle of an epoch (likely), then after we restart we only run 1
epoch with the remaining number of steps in that epoch, then we
restart again and run the rest of the epochs in the normal number
of steps per epoch.

Note that new workers currently always start from step = 0.
We need a way to bootstrap new workers to the step that everyone
else is on.