Revert to MultiWorkerMirroredStrategy, fix seg fault
This commit reverts back to MultiWorkerMirroredStrategy, which
fixes the seg fault that happened on restart with MirroredStrategy.
Some investigation revealed that the seg fault was independent of
MPI or horovod. As of this commit you can add new workers while
running horovod.

There are a few big concerns though:
1. The steps across the workers are growing apart; is this really
synchronized training still? Are they actually updating the same
model?
2. The new worker's throughput is much lower compared to existing
workers (13 images/sec vs 60+)