Fix expanding multiple times in one restart
Previously we will hang if 2 workers join at once. This is
because the new workers don't know of each other and so only ever
call expand once. This commit fixes this by using MPI broadcast
to synchronize whether we should keep expanding or not. This is
a cleaner approach than the previous, because we no longer need
to do the synchronization through the autoscaling agent.