Fix hanging caused by race condition on epoch end
After finishing the last step of an epoch, we transition back to
RUNNING because we didn't receive any signal to restart. However,
we never wait for everyone to be RUNNING, so some workers may
rush ahead and start syncing cluster spec while others are stuck
behind the barrier for (NOT_)READY_TO_RESTART.

This commit does two things:
(1) Add a synchronization barrier for RUNNING at the end of each
    step, and
(2) Add READY_TO_SYNC as one of the next states of RUNNING

Note that (2) is necessary because otherwise one worker may rush
ahead and transition to READY_TO_SYNC while others are stuck at
the synchronization barrier for RUNNING (adding in (1)).