Always save variables when restarting
This fixes a bug where we actually discard all our variables if
we restart WITHOUT external signal. We do this when finishing
an incomplete epoch. The right time to save variables then is
at the end of `model.fit`, not at the end of a batch when we
don't know yet if there will be more batches in this epoch.