Fix join during initialization race condition
Previously, if a worker joins while existing workers are syncing
cluster specs, then the resulting pending_cluster_specs may vary
across the cluster, resulting in some workers wanting to restart
and others not, and the job hangs.

This commit adds a safety check when the master server handles
join request: if the request is received during initialization,
the master server simply rejects the request and puts the burden
on the new worker to keep trying until master server is done with
initialization.