Fix join during initialization round 2
The previous commit did not completely fix the issue. The real
problem is when an epoch ends, we never transition to RESTARTING
even when we restart. Then, the master server can go ahead and
apply its pending cluster spec before handling the join request.
If everyone does the same then the new worker will be left out
of the cluster spec sync.

However, because mpi_spawned_communicators is not empty, the
existing servers will try to expand their communicators. This
will hang forever because the new worker was still stuck in the
cluster spec sync phase.

This commit ensures that
  (1) We will always be RESTARTING if we're restarting, and
  (2) We will always be READY_TO_SYNC (i.e. not accepting further
      join requests) before we apply our pending cluster specs.