Fix join during initialization part 3
So, it turns out the REAL problem was we restarted without
waiting for the new worker to show up in our pending cluster
spec. However, as soon as we spawn it we will already try to
expand our MPI communicator the next time we initialize.

This commit fixes this in a more systematic way. From now on,
all restarts are triggered solely by the presence of a pending
cluster spec. We no longer restart if we finished an incomplete
epoch. This also saves us some unnecessary initialization.