Wait until spawned workers join before syncing
This fixes a race condition that happens when we spawn a worker
right before restarting. What used to happen was the new worker
would be stuck on waiting to join the cluster, while the existing
clusters would be stuck on expanding the communicator. This is
because the master server is already expecting two communicators
to join but its pending cluster spec is only expecting one worker.
Thus, the fix is to make the two agree on the number of new
workers.