De-synchronize step end sync from spawning workers
If we sync at the same time as we spawn workers, then often we'll
end up missing the opportunity to add the spawned worker until
the next time we sync. For example, with cifar10 we sync every
10 batches and spawn a worker every 10 batches. If the new worker
is not ready by the time we start the next batch, then we will
have to wait another 10 batches before the worker can be added.
This results in 2 workers added in 20 batches instead of 1 worker
added in 10 batches.

This commit fixes this issue by adding an offset of 1 step when
syncing on step end.