Refresh function instead of rebuilding entire graph
This commit is the first step towards a faster restart when
changing the size of the horovod communicator. Previously we
always rebuilt the entire graph after hvd.init, but this takes
a long time: we lose all the pre-processed images and we need
to rebuild the model. After this commit, we only need to create
a new function every time we restart while using the same ops
created in the beginning.

Note that restart is not much faster than before yet because
when we spawn new workers we still have to wait for the new
workers to start from scratch. The next commit will further
remove this overhead by delaying when the new workers join
the existing cluster.