Significantly cut down restart time
This commit significantly reduces the time needed to change the
size of the cluster. This builds on top of the result from the
previous commit, which allows us to change the horovod size
without rebuilding the entire graph. Now we further delay when
the new worker joins the cluster to after it has already built
the model and run one batch.

The basic idea is simple, but there were practical challenges
that needed to be resolved to truly cut down the restart time:

(1) Bootstrapping the number of steps and epochs to train
(2) Adjusting the batch size when cluster size increases
(3) Controlling how many workers we wait for when we initialize

One low hanging fruit in further reducing the restart time
remains: make spawning happen asynchronously. Currently we
spawn the workers one after another, and each call to MPI_SPAWN
takes 3 seconds for some reason.