Avoid rebuilding functions for faster restart
This commit allows us to skip rebuilding functions when cluster
membership changes. Before we call `model.fit` in a loop. Now the
call to `model.fit` itself will handle restarts internally and
will reuse the functions that take the most time to build. This
logic is handled on the tensorflow side. All we have to do here
is set the right (global) variables when restarting or joining.

This commit also adds a BufferedIterator that is used for
reusing preprocessed data and adjusting the batch size dynamically.
When a worker is added, local batch size falls. Instead of making
a brand new dataset, however, we simply reuse the existing
iterator and buffer any leftover values to ensure we always
supply the current number of samples to the execution function.

Previously, the restart time for cifar10 was close to 15-20s
and that for imagenet was 25-30 seconds. As of this commit,
the restart time for both is around 3-4 seconds.