Fix new worker directory already exists race condition
The log directory is determined by the rank of the new worker,
which we previously set to be num current workers + num spawned
workers. Since we now spawn workers asynchronously, we may
not update the "num spawned workers" part in time for the sum
to reflect the existence of the new worker.

This commit fixes it by simply keeping a counter of the number
of workers in the system, both joined and pending. This counter
will then be used as the rank of the worker spawned next.