Remove status barrier, much faster restarts
This commit replaces the remaining status barriers with MPI
barriers, leading to much faster restarts when there are many
processes. The previous solution turned out to be not scalable;
syncing cluster specs with 60 workers took 7 seconds. After this
commit that takes less than 1 second. It also simplfies the code
significantly; autoscaling statuses are now only for the new
worker to decide when to join and when to fetch various items
from existing workers.

Note that this commit uncovered the following race condition:
if workers A, B, and C were spawned in that order and B joins
before A and the existing cluster waits for B only, then the
whole system hangs. This commit adds a minimum number of steps
between restarts to workaround this issue. In the future we may
want to solve this more systematically, e.g. by tagging the
spawned communicator with its expected rank and wait specifically
for that rank during initialization.