Fix hanging caused by out of order worker joins
Before we only waited for N workers to join before expanding our
communicator. Unfortunately, in the event of workers joining out
of order, this can hang because the master server expects a
different communicator to participate in the expanding process.

This commit fixes this by keeping track of which spawned
communicator belongs to which rank, and rejecting workers who
are joining too early. Care must be taken to avoid deadlocks,
which occurred pervasively in the first two drafts of this commit.
In particular, we must ensure that holding `agent.spawn_lock` in
the join request handling code does not block the service for a
long time. More detail can be found in a comment in service.py.