Handle remove workers
There were a few things that needed to be done:
(1) MPI communicator needs to be downsized
(2) CUDA_VISIBLE_DEVICES assigned to the removed workers need
    to be freed, and
(3) BufferedIterator needs to be able to handle increases in
    local batch sizes

The only tricky problem was (3), and the solution is always use
the global batch size to define the returned tensor shape. One
concern with this approach is that if the global batch size is
large then there may be some additional memory pressure. It may
be fine in the end because we only pad the extra space with zeros.