Curve fitting: backtrack to see if we overshot
When stop conditions are met, we may have jumped too far in the
curve, i.e. overshot. When this happens, we scale down one worker
at a time to find the point when stop conditions were last not
met. Note that if we already have the data point for it then
we don't need to backtrack. E.g.

Num workers over time: 12, 24, 48 (stop), 47, 46, 45, 44 (stay)

Here we didn't have the data point for 47 workers, so we did
need to backtrack from 48 workers.