Implement straggler detection
A process is considered a candidate straggler if its computation
time is more than two standard deviations above the mean. If a
process stays a candidate straggler for more than N = 100 rounds
in a row then it is considered a real straggler.

Note that this commit does not include the code to replace the
stragglers yet. This will come in the next commit.