Make checkpoint-restart work with new schedule calback
With the new schedule callback, checkpoint-restart mode no longer
worked because the throughputs from before restarting are missing.
This commit saves these throughput averages and restores them
across restarts.

Additionally, we fix a significant bug in the way we compact the
throughputs. The computed averages were much lower than what they
should be. This is no longer a problem.