New straggler detection: more intuitive, less conservative
The new straggler detection computes each worker's throughput
as a fraction of the median throughput in that batch. We then
store the EWMA of these throughput fractions. A worker is
considered a straggler if its averaged throughput fraction
falls below a certain value (e.g. 0.8).

This has several advantages over the previous approach:
- We score how likely a worker is going to be a straggler
  instead of classify them as either 0 or 1
- We are less conservative than the previous approach; it's
  difficult for a worker to be in the top (bottom) 5th percentile
  for 100 batches in a row
- We handle gaps in straggler behavior; just because a worker
  stopped being a straggler for 1 batch doesn't mean it's no
  longer a straggler

The results are a more intuitive threshold (minimum allowable
fraction of median throughput) and a more robust detection
algorithm.