Remove curve fitting, add decrease logic in linear mode
Curve fitting was unreliable. From now on, we will simply add
and remove a fixed number of workers every N batches. This commit
also adds the optimization to wait for new workers to join while
running instead of during initialization, which saves us a few
seconds when restarting with multiple spawned workers.