Make sure we have enough information before settling
This is another attempt at solving different starting points
converging to different ending points. There were cases when
scaling up converges to slightly lower numbers of workers than
scaling down. This is likely because the workers are already
warmed up from scaling down so the throughputs are closer to
the steady state. This commit adds the minimum number of batches
we need for k and k+n workers before settling on k.

Note: this commit covered another problem, where if we remove
and spawn too many workers we'll get MPI_SPAWN_ERROR, because
the old workers are never removed. The next commit will fix this
by reusing the old workers. This is the better approach in
practice anyway, because until we get to steady state we want
to keep them around just in case.