In autoscaling mode, let master spawn all other workers
We need this functionality to make re-adding removed workers work.
Currently we can only re-add workers that are spawned through MPI.
The problem before was that workers that were launched in the same
world as the master could not be re-added. This commit adds some
startup overhead (1 or 2 minutes) but the runtime performance is
not affected. Also, this should only affect autoscaling mode.