Allow global batch size to rise up to a limit
This commit entails a large change in how the system works.
Previously we always fixed global batch size, but this does
not scale well for ImageNet on GPUs, for example. For workloads
with well known convergence characteristics, the user should be
able to set an upper bound for global batch size if he wishes
to avoid compromising model quality. The system can still freely
scale before this maximum is hit. After the maximum is hit, the
system will revert to previous behavior of fixing the global
batch size while scaling up.

This change makes the system a lot more practical.