Add Horovod integration
When Horovod is enabled, we use the mirrored distribution strategy
instead of the multi-worker mirrored strategy. Under this scheme,
there is one worker on each machine. We still use tensorflow's
distribution strategy to synchronize gradients across devices on
the same machine, but we will use Horovod to further synchronize
across the workers.