Fix bert fine-tuning, shard by input context
We did this for ResNet on ImageNet/CIFAR-10 but never did it
for BERT fine-tuning. Consequently this workload failed before
this commit, complaining that the number of input iterators
(which was always 1) didn't match the number of virtual nodes
per device. This is simply fixed by adding the proper sharding
code as we have done in the ResNet case.