Add helper callback to limit checkpoints kept
This commit adds the --num_checkpoints_to_keep flag for resnet
workloads that controls the number of checkpoints to keep. This
flag is read only if --enable_checkpoint_and_export is set.
This commit also adds checkpointing functionality to the cifar10
workload, and put the checkpoints in the log dir for convenience
when running in a docker environment.