Enable virtual node processing in BERT workloads
As of this commit, BERT pretraining, glue finetuning, and squad
finetuning all work with virtual nodes. This commit also cleans
up the run scripts by abstracting common flags. Because BERT
pretraining and squad finetuning relies on the custom run loop
implemented in this repo (instead of the keras run loop in the
TensorFlow repo), we needed to re-implement virtual node
processing here, though much of the code can simply be copied
over from the TensorFlow implementation.