Add support for elasticity
The two main parts of this commit are (1) the resize mechanism,
and (2) synchronizing resize across workers. Part (1) relies on
replacing the underlying extended distributed strategy with a
new one built using a new TF_CONFIG, then using the strategy's
new group key in the tensorflow graph. This commit implements
only the first half of this mechanism, and the other half is
implemented in the tensorflow repo.

Part (2) synchronizes resize across workers with MPI. The master
broadcasts its resize state at the beginning of each batch. The
resize state can be set by the user in a python shell through
`client.set_num_workers(n)`. If the resize state is set, then
the above resize mechanism will be triggered for all workers.

One major limitation of this implementation is removed workers
do not actually exit. This is due to complications in adding
new workers that were not part of the original cluster. This
feature may be added in a future commit.