Implement elasticity using Horovod
This addresses the limitation of the previous implementation,
which did not rely on Horovod, that new workers cannot be added
to an existing cluster. With Horovod, each process uses
MirroredStrategy on its own and is connected with each other
through an allreduce function that uses the global MPI
communicator that involves all workers in the cluster. Resizing
the cluster shrinks or expands this MPI communicator accordingly
and reinitializes Horovod to use the new communicator.