Elasticity at the granularity of GPUs, not machines
As of this commit, each process will be assigned a single GPU
in elasticity mode. The master will maintain a global map of all
GPUs assignments in the system, and will use this information to
set CUDA_VISIBLE_DEVICES when spawning processes. This allows us
to resize at the granularity of GPUs instead of nodes.

One major change in this commit was spawning each new process
individually instead of spawning all of them at once. This
change was needed because the MPI bulk spawn API used previously
did not allow the master to launch specific ranks on specific
machines, which is needed to maintain the CUDA_VISIBLE_DEVICES
mapping.