Implement weighted fair sharing + elasticity
This commit implements the bulk of the logic of the scheduler.
Jobs are allocated resources proportional to their priorities.
If a new job comes in with a high priority, it will partially
preempt existing jobs with lower priorities, triggering a scale
down for the affected jobs. Similarly, when a job exits, its
resources will be released back into the pool and distributed
among remaining jobs in the system, triggering a scale up for
these jobs.