Scheduler: fix ignored resize to 1 worker request
The root cause of this issue was the scheduler first checked if
the job already has the target number of workers, and if so,
drop the request. This is wrong, however, if the scheduler
queried the job size before the initial set of workers has
joined, because the scheduler will mistakenly think the job
already has 1 worker and there is no need to send the resize
request.

Instead, we should always send the request and let the job take
care of it. If the job is waiting for spawned workers, it will
throw an exception to inform the scheduler to try again later.
This commit removes the aforementioned check to fix this issue.

After removing this check, we don't need the `get_num_workers`
API on the elasticity callback anymore, so this commit also
removes this API.