Scheduler: fix job env var overridden by system
Before this commit, if the user set ENABLE_ELASTICITY=false in
the shell and ran scheduler.py, the jobs will fail because the
system env vars will overwrite the job env vars, and elasticity
will be disabled. Instead, we should separate the two and update
the job env vars at the very end so they will not be overwritten.