Fix wrong batch size in elasticity mode, led to OOM
In non-elasticity mode, we relied on tensorflow's
experimental_distribute_dataset to split our batch further into
the number of devices. In elasticity mode, each GPU has only 1
process, so this doesn't actually divide our batch by the number
of devices. As a result, the total batch size is divided by only
the number of virtual nodes, and the virtual node size in each
batch often exceeded the memory of an individual GPU.

This commit fixes it by ensuring that, in elasticity mode, we
divide each batch by the _total_ number of virtual nodes rather
than the number of virtual nodes for each device.